Solventsthreads must >= partInCompound. Lots of working. i am drowning

+++ Quantom/Quantom/Engine.cuh
@@ -32,7 +32,7 @@ public:
 		LIMAENG::applyHyperpos(pos1, pos2);
 
 
-		double vel = (*pos1 - *pos2).len() * 0.5l / dt;
+		float vel = (*pos1 - *pos2).len() * 0.5f / dt;
 		float kinE = 0.5 * mass * vel * vel;
 		return kinE;
 	}
